# Add functions to decompose einsum summation