--- a/tests/test_einsum_optimizers.py
+++ b/tests/test_einsum_optimizers.py
@@ -109,6 +109,7 @@ class EinsumOptimizerTests(Tf2OnnxBackendTestBase):
         self.common_einsum('obk,ijk->ioj', operands=[f, g],
                            catch_errors=False)
 
+    @check_opset_min_version(11, "Einsum")
     def test_np_test_exp(self):
         self.common_einsum('iij,jk->ik', catch_errors=True)