--- a/tf2onnx/graph.py
+++ b/tf2onnx/graph.py
@@ -1004,7 +1004,11 @@ class Graph(object):
         def _push_stack(stack, node, in_stack):
             stack.append(node)
             if node in in_stack:
-                raise ValueError('Graph has cycles, node=' + ops[node].name)
+                import pprint
+                pprint.pprint(in_stack)
+                pprint.pprint(stack)
+                pprint.pprint(ops)
+                raise ValueError('Graph has cycles, node.name=%r.' % ops[node].name)
             in_stack[node] = True
 
         def _get_unvisited_child(g, node, not_visited):