This code decompose an einsum equation into an ONNX graph. There are still some limitations:

* expression `...`  is not supported yet
* diagonal is not suppoted (example `ii->i`)
* output must be specified (no expression such as `ab,bc`
* the graph is not always optimized (consecutive transpose are still remaining)

*Note:*

operator MatMul is not the most efficient way to decompose the einsum summation (MatMul = `ijo,iok->ijk`) which is equivalent to i matrix multiplications. Operator Gemm is more appropriate to do less transpose (Gemm = `jo,ko->jk`) but multiple Gemm cannot be done with the current implementation. Gemm could still be used to do `ajo,bko->abjk` by merging the two first dimensions on each input.

Benchmark (on a Windows Surface):

**Model small_bert_bert_uncased_L-8_H-512_A-8_2**

Goes from 0.06657s to 0.05109s (~20% faster)

**Model expert_bert_wiki_books**

Goes from 0.77959s to 0.67123s (~14% faster).

**Inputs:**

same inputs for both models 

```
def get_input():
    return {"input_type_ids": get_small_rand_int32([1, 512]),
                "input_mask": get_ones_int32([1, 512]),
                "input_word_ids": get_zeros_then_ones([1, 512])}
```

**Example for equation `bsnh,btnh->bnts`**

![image](https://user-images.githubusercontent.com/22452781/116728204-ea649f80-a9e5-11eb-9442-371a34427115.png)